Add support for detailed API calls and result POSTing
I was informed that more details about the submitter and other metadata is
available in the Series API if called with a special query parameter.  This
has been added, since it's useful and there's no other way to retreive it
in the API at present.

In addition, if the patch failed to merge results are POSTed to the API.
It is pointing to my personal Patchwork setup which is configured to not
mail the author or the mailing list, but just to email me for testing.
Be careful that you don't use a Patchwork instance that can send real email
while testing.

Signed-off-by: Russell Currey <ruscur@russell.cc>