Support specification of Jenkins job variables in config file
Jobs are now specified as a TOML array of tables. Each table is a job, with
the field "job" for the job name, "remote" for the Jenkins configuration
variable corresponding to the git remote, "branch" for the Jenkins
configuration variable corresponding to the git branch. Any other field is
treated as a variable to be passed straight to Jenkins.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>