git: move apply patch code into separate, generic function
Create a new function, git::apply_patch(), that takes a Repository and a
path to an mbox file and attempts to apply the patch.

As part of this refactoring, test_patch() now returns immediately when the
patch fails to apply. Right now, this breaks posting of failure results to
patchwork - but test result posting is pretty broken anyway and will be
fixed soon.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>