Initial support for testing from multiple base branches
Many large projects have a number of different branches, such as "master",
"next" and "stable". A submitted patch may be intended to be applied on top
of any one of these branches, or possibly for multiple branches.

To allow testing on multiple base branches, rather than specifying a single
base branch, we now specify a list of base branches in the project
configuration.

By default, snowpatch will attempt to apply patches on every base branch
and trigger a Jenkins job for every branch. By setting "test_all_branches =
false" in the project configuration, snowpatch will go through the listed
base branches in order and trigger a Jenkins job on the first branch on
which the patch is successfully applied.

This support needs some cleanup and some more work, but this is a start.

Closes: #2 ("Multiple branch support")
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Russell Currey <ruscur@russell.cc>