Use patchwork "linkname" instead of "name"
Patchwork gives us both the name of the project, and the linkname.  For
example, on patchwork.ozlabs.org the name of skiboot is "skiboot firmware
development", and the linkname (which appears in the URL) is "skiboot".
It makes more sense for snowpatch to use the latter rather than something
verbose and descriptive.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>