Don't retest patches that have already been tested
Currently when snowpatch is running in endless/daemon mode, once it's
tested all the series it will go back and do it all again.  This is a
waste of time and resources, so let's not do that.

If Jenkins tests are once again parallelised, and the Patchwork polling
interval is set to something low, there's a chance that the test state
of a series will not be reported and that something will still be tested
twice.  We can deal with that when/if it happens.

Closes: #12 ("'Endless' mode retests series")
Signed-off-by: Russell Currey <ruscur@russell.cc>
[ajd: rebased patch on top of 0650293]
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>