Include branch name in test results sent to Patchwork
If a project is configured to test on multiple branches and a patch applies
to more than one base branch, only one result for each Jenkins job will
appear in Patchwork since the test name is the same.

Reformat the Patchwork test names to be of the format "branch/test_name" so
we can distinguish between different base branches.

Closes: #25 ("Test names sent to Patchwork need to differ per branch")
Reported-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Russell Currey <ruscur@russell.cc>