HTTP proxy support
With the new HTTP proxy support in hyper 0.9.2, we can finally include it
in snowpatch.

Enable snowpatch to use HTTP proxies through the http_proxy environment
variable, as seems to be standard.

Hyper does not yet support HTTPS proxies.  In addition, the Jenkins code
does not yet support using the main hyper Client, so the proxy will not
be used.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>