Enable env_logger logging, use logging macros
Initialise the env_logger logging crate. By default, we display
messages of the level "info" or higher, and we use the default "RUST_LOG"
environment variable for user-specified logging settings.

Replace existing print messages with appropriate levels of logging, and add
a couple of new messages. There is still a lot more we should log, but
this is a decent start.

Signed-off-by: Russell Currey <ruscur@russell.cc>
[ajd: reword commit message]
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>