settings: improve parsing of jobs
Job entries in the config file are a mixture of parameters which should be
passed straight through to Jenkins, and flags that have meaning to
snowpatch.

I don't really feel like splitting these two things apart in the config, as
we already have enough levels of nesting. Instead, add our own deserialize
method to do this more intelligently. This also serves as an example of how
to do more advanced parsing such that we can reject invalid values at parse
time.

Thanks-to: David Tolnay <dtolnay@gmail.com> # help with boilerplate code
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Acked-by: Russell Currey <ruscur@russell.cc>