Upstream Patchwork support
Implement support for the patch and series models in upstream Patchwork.

In order to support this different API, snowpatch has been re-architected
around having distinct concepts of patches and series.  Instead of treating
every patch as a series, we instead operate on every patch, and for patches
that are in the middle of a series, we apply all of its dependencies before
testing.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Co-authored-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
[ajd: rebase on serde changes, lots of fixes, token authentication]
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>