Remove jenkins_test and explicit [[bin]] section
With the rust updates, either we need to remove the [[bin]] section
in Cargo.toml to allow auto-inferred binaries, or manually specify
where jenkins_test is.  While we could do that, I think it has bitrot,
it has hardcoded stuff in it and isn't really necessary when we can
test Jenkins stuff pretty easily through snowpatch itself now.

So, just get rid of it.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>