Convert from hyper to reqwest
There's a bit going on here but it's hard to separate without breaking bisect.

Firstly, convert from hyper to reqwest.  We have no need for asynchronous HTTP
and the new hyper API isn't suitable for our needs, whereas reqwest is pretty
close to what we originally used hyper for, except somewhat simpler.

Next, update the rest of the dependencies to their latest major versions.
This *just compiled*.

Lastly, change the Jenkins HTTP helpers to be a bit more like the Patchwork
ones.  This worked around an issue with different return types and seemed like
the simplest solution.

Signed-off-by: Russell Currey <ruscur@russell.cc>