Require project argument to be specified for daemon mode
So this might seem out of place, but bear with me here.

Right now, the way that we check for patches is to get the latest
patches "page" from the API.  For each of them, we see if the project
the patch is from matches one we have settings for, and if so we
test it.  We exhaust the patch list page we got, and then we get a
new one.

There are lots of problems with this.  The first one is that if we're
testing lots of patches and the Patchwork instance is high traffic,
we risk missing patches that fall to page 2 or beyond in that time.
We currently don't have a solution for this.

The next issue is that when I originally implemented this, filtering
by project didn't work properly in the patches API, making it
difficult to track all patches, especially for niche projects with
less activity (since running into one patch of a series will lead
to them all being tested, even if snowpatch doesn't find them all
individually in the patch list).

This is no longer the case however, and we can filter by project
on the patch list and actually get a full page of patches in that
project.  As a result, until both the following conditions are true:

 - patchwork has a suitable events API and/or better pagination
 - snowpatch has a proper threading model for distributed work

Limit snowpatch to one project at a time to minimise the chance of
missing patches, and users should just spin up more than one
instance if they want to test multiple projects at once.

Signed-off-by: Russell Currey <ruscur@russell.cc>