Fix --series option
Somewhere along the way, we completely broke the --series option and failed
to notice.

It appears we try to retrieve a Project over the Patchwork API and pass an
entirely invalid URL. As it turns out we can just use the Project already
retrieved as part of the Patch that we already retrieve, so use that
instead.

(This is why we need tests, which we ironically don't yet have...)

Reported-by: Russell Currey <ruscur@russell.cc>
Closes: #43 ("Fix --series")
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>