Support pagination when using --count
Add support for pagination when using --count. We extract the Link header
from the Patchwork response to find the correct URL for the next page.

The pagination is potentially racy, as new patches could come in and
disturb the ordering, but this is a limitation on the Patchwork side at
present and is unlikely to be a significant problem.

Closes: #17 ("Enable the ability to go to page 2 (and beyond) of Patchwork results")
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Russell Currey <ruscur@russell.cc>