Update to reqwest 0.9 and temporarily disable pagination
reqwest 0.8 uses an older version of hyper which uses an older version
of the openssl crate, which means that snowpatch can't build on any
particularly modern distribution.  Updating reqwest to 0.9 however is
quite the ordeal, because it now uses hyper 0.12 which no longer has
stringly typed header values, which isn't too much hassle...

except for the following:

 - authorization, which we now have to do manually, which is
   particularly annoying for the Basic Auth case

 - pagination using the Link header, which is now a significant
   undertaking to parse

I am opting on the side of disabling pagination until some kind of
helper is released, or we put in the effort to write one ourselves.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>