Rewrite settings parsing to avoid panic
Change parse() to return a Result, using the question mark operator to
catch errors and return early.

Use map_err() to turn the low-level errors into something human
readable. We could define our own error type but that seems like
overkill for this.

Rework the tests to cope. For the two failing cases that's a bit ugly
because we have to map the Ok() to Err() and vice versa.

There are now no panic()s or unwrap()s in settings.rs!

Error messages seen by the user look like, eg:
  Error: Unable to parse config file: missing field `git`
  Error: Unable to open config file: Permission denied (os error 13)
  Error: Unable to open config file: No such file or directory (os error 2)

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>