jenkins: Get started on fixing error handling
We use unwrap() a lot. We should do less of that.

Get rid of a lot of the unwrap()s in the Jenkins code and return Result
types with errors instead.

We haven't completely gotten rid of unwrap(), there's still some bits which
are a bit gross, and on the main side we still just unwrap rather than
handling the error properly.

At some point in the future, we might switch to using an external crate for
the Jenkins API itself to make life a bit easier.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Russell Currey <ruscur@russell.cc>